+++ src/main/java/DAL/Dao/Interfaces/IMarkersDao.java
@@ -1,5 +1,7 @@
 package DAL.Dao.Interfaces;
 
+import DAL.Dao.MarkersDao;
+import DAL.EntityDataBase.MarkersDto;
 import Model.Markers;
 import com.google.maps.model.LatLng;
 
@@ -8,8 +10,6 @@ import java.util.List;
 /**
  * Created by Ignacio on 11/06/2017.
  */
-public interface IMarkersDao {
-    List<Markers> getActiveMarkers(LatLng latLng);
-    List<Markers> getNearMarkers(LatLng latLng);
+public interface IMarkersDao extends IBasicDao<MarkersDto> {
 
 }
